+++ packages/@coorpacademy-components/src/organism/brand-analytics/style.css
@@ -1,0 +1,45 @@
+@value breakpoints: "../../variables/breakpoints.css";
+@value mobile from breakpoints;
+@value tablet from breakpoints;
+@value desktop from breakpoints;
+@value colors: "../../variables/colors.css";
+@value dark from colors;
+@value light from colors;
+
+.dashboardContent {
+  font-family: "Open Sans";
+  width: 100%;
+  text-align: center;
+  margin: 0 20px 20px 0px;
+  display: flex;
+  flex-direction: column;  
+  min-height: 700px;
+}
+
+.dashboardSelection {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+}
+
+.dashboardTitle {
+  font-size: 24px;
+  color: dark;
+  border: 1px solid light;
+  padding: 10px;
+  flex-grow: 0;
+}
+
+.dashboardIframe {
+  width: 100%;
+  height: 100%;
+  flex-grow: 1;
+}
+
+.dashboardNoSelection {
+  font-size: 20px;
+  font-weight: bold;
+  padding: 40px;
+}
+
+/* TODO: mobile rendering */
