+++ packages/@coorpacademy-components/src/organism/brand-analytics/test/fixtures/loading.js
