+++ packages/@coorpacademy-components/src/organism/brand-analytics/test/fixtures/selected-loading.js
