--- packages/@coorpacademy-components/src/organism/brand-home/style.css
@@ -1,23 +1,0 @@
-@value colors: "../../variables/colors.css";
-@value dark from colors;
-
-.home {
-flex: 1;
-display: flex;
-justify-content: center;
-align-content: center;
-align-items: center;
-}
-.homeText {
-  width: 611px;
-  height: 58px;
-  font-family: Gilroy;
-  font-size: 48px;
-  font-weight: bold;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: normal;
-  text-align: center;
-  color: #111117;
-}
