--- packages/@coorpacademy-components/src/organism/brand-home/test/fixtures/default.js
@@ -1,5 +1,0 @@
-export default {
-  props: {
-    text: 'Welcome to coorpmanager'
-  }
-};
