+++ packages/@coorpacademy-components/src/template/back-office/brand-update/test/fixtures/dashboards-analytics.js
@@ -1,5 +1,5 @@
 import {defaultsDeep} from 'lodash/fp';
-import BrandDashboard from '../../../../../organism/brand-dashboard/test/fixtures/selected';
+import BrandDashboard from '../../../../../organism/brand-analytics/test/fixtures/selected';
 import Default, {items} from './default';
 
 const {props} = Default;
@@ -8,7 +8,7 @@ export default {
   props: defaultsDeep(props, {
     items,
     content: defaultsDeep(BrandDashboard.props, {
-      type: 'dashboard',
+      type: 'analytics-dashboards',
       header: {}
     })
   })
