+++ packages/@coorpacademy-components/src/template/back-office/brand-update/test/fixtures/home.js
@@ -6,7 +6,7 @@ const {props} = Default;
 export default {
   props: defaultsDeep(props, {
     content: {
-      type: 'home',
+      type: 'dashoard',
       text: 'Welcome to coorpmanager'
     }
   })
